update readme etc
